requirements.txt
google-api-python-client==2.108.0
pandas==2.1.3
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
yt-dlp==2023.11.16
groq==0.4.2
pytz==2023.3
requests==2.31.0
